Change database to LocalDB (no SQLExpress any more); change date validation; change signalr and mvc references
NB SignalR stuff does not work. Knockout stuff in Products View is OK, though a bit buggy.

+++ src/KnockoutFirstKickOfTheCat/Controllers/ProductController.cs
@@ -141,12 +141,11 @@ namespace KnockoutFirstKickOfTheCat.Controllers
 
             return RedirectToAction("Index");
         }
-
         
         private void DeleteProduct(Product product)
         {
             Product repoProduct = db.Products.Find(product.Id);
-            db.Products.Remove(repoProduct);
+            if (repoProduct != null) db.Products.Remove(repoProduct);
             db.SaveChanges();       
         }
